% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuster_indices_takuzu.R
\name{ajuster_indices_takuzu}
\alias{ajuster_indices_takuzu}
\title{Ajuste les cases d'une matrice en fonction des erreurs du joueur}
\source{
"generate_valid_matrix.R"
}
\usage{
ajuster_indices_takuzu(matrice_base, matrice_premplie)
}
\arguments{
\item{matrice_base}{Une matrice carree representant la matrice de base valide du jeu Takuzu, generee par une fonction comme \code{generate_valid_matrix}.}

\item{matrice_premplie}{Une matrice carree representant l'etat actuel du jeu apres que le joueur ait rempli certaines cases, avec des cases vides (\code{NA}).}
}
\value{
Une matrice de meme taille que \code{matrice_premplie}, avec les cases vides (\code{NA}) remplacees par les valeurs correspondantes de \code{matrice_base}.
}
\description{
Cette fonction permet de comparer une matrice de jeu pre-remplie par le joueur avec une matrice de base valide.
Les cases vides (\code{NA}) dans la matrice pre-remplie sont remplacees par les valeurs correspondantes de la matrice de base.
Cela permet d'ajuster la grille du jeu en fonction des erreurs du joueur et de completer la grille avec les bonnes valeurs.
}
\examples{
# Generer une matrice valide de Takuzu
base_matrice <- generate_valid_matrix(6)

# Une matrice pre-remplie par le joueur avec des erreurs (NA)
matrice_joueur <- base_matrice
matrice_joueur[1,1] <- NA
matrice_joueur[2,2] <- NA

# Ajuster les erreurs du joueur
matrice_corrigee <- ajuster_indices_takuzu(base_matrice, matrice_joueur)

}
