% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_valid_matrix.R
\name{generate_valid_matrix}
\alias{generate_valid_matrix}
\title{Génère une matrice valide pour le jeu Takuzu}
\source{
"generate_valid_row.R"

"rotate_vector.R"
}
\usage{
generate_valid_matrix(taille)
}
\arguments{
\item{taille}{Un entier représentant la taille de la matrice carrée à générer (par exemple, 8 pour une matrice de 8x8).}
}
\value{
Une matrice carrée de taille \code{taille}x\code{taille} remplie de 0 et 1, où chaque ligne est une rotation circulaire de la première ligne.
}
\description{
Cette fonction génère une matrice carrée valide pour le jeu Takuzu. Elle utilise la fonction \code{generate_valid_row} pour
générer une ligne valide, puis crée une matrice en déplaçant cette ligne de manière circulaire pour chaque ligne suivante.
Les règles du jeu Takuzu sont respectées : chaque ligne contient autant de 0 que de 1, et les lignes sont des rotations
circulaires de la ligne générée initialement.
}
\examples{
# Générer une matrice valide de taille 6x6
generate_valid_matrix(6)

}
