% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_valid_row.R
\name{generate_valid_row}
\alias{generate_valid_row}
\title{Génère une ligne valide pour le jeu Takuzu}
\usage{
generate_valid_row(taille)
}
\arguments{
\item{taille}{Un entier représentant la taille de la ligne à générer.}
}
\value{
Un vecteur logique de taille \code{taille} contenant des valeurs 0 et 1.
}
\description{
Cette fonction génère une ligne valide pour le jeu Takuzu en respectant les règles de base :
\itemize{
\item Chaque ligne doit être remplie de manière équilibrée avec des 0 et des 1,
\item Il ne doit pas y avoir trois chiffres consécutifs identiques dans la ligne.
}
}
\examples{
# Génère une ligne valide de taille 8
generate_valid_row(8)

}
