% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_vector.R
\name{rotate_vector}
\alias{rotate_vector}
\title{Effectue une rotation circulaire d'un vecteur}
\usage{
rotate_vector(x, shift)
}
\arguments{
\item{x}{Un vecteur de valeurs (par exemple, numériques, logiques, ou de caractères) à faire tourner.}

\item{shift}{Un entier représentant le nombre de positions à décaler. \code{shift} est toujours positif, et la rotation se fait vers la droite.}
}
\value{
Un vecteur de même type et longueur que \code{x}, avec les éléments tournés.
}
\description{
Cette fonction effectue une rotation circulaire d'un vecteur en déplaçant ses éléments vers la droite ou la gauche,
selon le nombre de positions spécifié par l'argument \code{shift}.
Les éléments sont déplacés et réintégrés à l'extrémité opposée de la séquence.
}
\examples{
# Effectuer une rotation à droite de 2 positions
rotate_vector(c(1, 2, 3, 4, 5), 2)

}
