% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Serveur pour l'application Shiny Takuzu}
\source{
"generate_valid_row.R"

"rotate_vector.R"

"generate_valid_matrix.R"

"creer_matrice_takuzu.R"

"verifier_grille.R"

"ajuster_indices_takuzu.R"

"ui.R"
}
\usage{
server(input, output, session)
}
\arguments{
\item{input, }{output, session Parametres standards de Shiny. Ces objets sont utilises pour reagir aux evenements de l'utilisateur et mettre a jour l'interface.}

\item{output}{L'objet output pour l'affichage dans l'UI.}

\item{session}{L'objet session pour la gestion de l'etat de l'application.}
}
\value{
Une application Shiny fonctionnelle.
}
\description{
Cette fonction gere le serveur de l'application Takuzu, y compris la logique de generation et d'affichage de la grille, l'interaction avec l'utilisateur, la verification de la grille, la gestion des indices et l'affichage du resultat.
}
