% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_matrice_takuzu.R
\name{creer_matrice_takuzu}
\alias{creer_matrice_takuzu}
\title{Cree une matrice de jeu Takuzu avec des cases vides selon la difficulte}
\usage{
creer_matrice_takuzu(matrice_base, difficulte = "facile")
}
\arguments{
\item{matrice_base}{Une matrice carree generee par \code{generate_valid_matrix}, representant la base valide du jeu.}

\item{difficulte}{Une chaine de caracteres specifiant le niveau de difficulte. Les options sont :
\itemize{
\item "facile" : ajoute environ 10\% de cases manquantes.
\item "avancee" : ajoute environ 25\% de cases manquantes.
\item "difficile" : ajoute environ 35\% de cases manquantes.
}
Par defaut, la difficulte est "facile".}
}
\value{
Une matrice de meme taille que \code{matrice_base}, avec des cases vides (\code{NA}) reparties selon la difficulte choisie.
}
\description{
Cette fonction genere une matrice de jeu Takuzu en fonction de la difficulte choisie. Elle utilise une matrice de base valide (par exemple, generee par \code{generate_valid_matrix})
et remplace certaines cases par des \code{NA} pour rendre le jeu plus difficile selon le niveau specifie. Le taux de cases manquantes varie selon la difficulte.
}
\examples{
# Creer une matrice de Takuzu facile
base_matrice <- generate_valid_matrix(6)
matrice_facile <- creer_matrice_takuzu(base_matrice, difficulte = "facile")

# Creer une matrice de Takuzu difficile
matrice_difficile <- creer_matrice_takuzu(base_matrice, difficulte = "difficile")

}
