% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Serveur pour l'application Shiny Takuzu}
\source{
"generate_valid_row.R"

"rotate_vector.R"

"generate_valid_matrix.R"

"creer_matrice_takuzu.R"

"verifier_grille.R"

"ajuster_indices_takuzu.R"

"ui.R"
}
\usage{
server(input, output, session)
}
\arguments{
\item{input, output, session}{Paramètres standards de Shiny. Ces objets sont utilisés pour réagir aux événements de l'utilisateur et mettre à jour l'interface.}
}
\value{
Une application Shiny fonctionnelle.
}
\description{
Cette fonction gère le serveur de l'application Takuzu, y compris la logique de génération et d'affichage de la grille, l'interaction avec l'utilisateur, la vérification de la grille, la gestion des indices et l'affichage du résultat.
}
